rss: move IRssObserver into seperate header file
